# docs/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Varka Backend API
  version: 0.0.1
  description: REST API for authentication, projects, CIM/PIM/PSM models, transformations, artifacts, modeling metadata, and diagram layout.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - AuthToken: []
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects visible to the current user
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveProjectRequest"
      responses:
        "200":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/projects/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      summary: Get a project
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Projects]
      summary: Update a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Projects]
      summary: Delete a project
      responses:
        "200":
          description: Project deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/projects/{id}/download:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      summary: Download project ZIP
      responses:
        "200":
          description: Project ZIP archive
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/projects/{id}/members:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      summary: List project members
      responses:
        "200":
          description: Project members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/projects/{id}/invite:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      tags: [Projects]
      summary: Invite an existing user to a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteRequest"
      responses:
        "200":
          description: Added member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/projects/{id}/members/{userId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - name: userId
        in: path
        required: true
        schema:
          type: string
    put:
      tags: [Projects]
      summary: Update a project member role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRoleRequest"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Projects]
      summary: Revoke project access
      responses:
        "200":
          description: Access revoked
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/impact/artifact/{artifactId}:
    get:
      tags: [Impact]
      summary: Analyze generated artifact lineage
      parameters:
        - name: artifactId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artifact metadata and ordered upstream model ancestors
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    AuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Unauthorized:
      description: Missing, invalid, or expired auth session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Forbidden:
      description: Authenticated user does not have permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    NotFound:
      description: Record or file not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
  schemas:
    ApiErrorResponse:
      type: object
      required: [message, status, timestamp, issues, errorId]
      properties:
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        issues:
          type: array
          items: {}
        errorId:
          type: string
          description: Correlation id for support and log lookup
...
    SaveProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        activeModelIds:
          type: object
          additionalProperties:
            type: string
          description: Map of active model level or view keys to model ids.
    InviteRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          maxLength: 48
    UpdateMemberRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          maxLength: 48
    ProjectRecord:
      type: object
      required: [id, name, ownerUserId, activeModelIds, members, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        ownerUserId:
          type: string
        activeModelIds:
          type: object
          additionalProperties:
            type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMember"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectMember:
      type: object
      required: [userId, email, displayName, role, addedAt]
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
        addedAt:
          type: string
          format: date-time
...